LayoutTests/platform/wpe/imported/w3c/web-platform-tests/dom/events/Event-timestamp-high-resolution-expected.txt

CONSOLE MESSAGE: line 7: ReferenceError: Can't find variable: GamepadEvent

FAIL Untitled ReferenceError: Can't find variable: GamepadEvent
